Core: Emit tunnel snapshots from SimpleConnectionDaemon (#401)
Submit periodical daemon snapshots to the TunnelController, where they
can be forwarded to the native layers.

This design suits the Android activity/service model better, where the
service may broadcast any sort of data to the application. This is not
the case on Apple systems, where the tunnel events are limited to the
Network Extension metadata (status, isEnabled, onDemand).

Apple:

- Snapshots are collected by the app directly
- In the tunnel provider, the `SimpleConnectionDaemon` connection writes
to an environment typically backed by a `UserDefaults` in the App Group
that the app can also read from
- In the app, `NETunnelStrategy` observes a set of tunnel managers
(`NETunnelProviderManager`)
- In the app, `Tunnel` produces the snapshots from the strategy managers
and the `UserDefaults` environments where the tunnel writes to,
periodically fetched

Android:

- Snapshots are collected by the service and broadcast to the app
- In the tunnel service, the `SimpleConnectionDaemon` connection
produces full snapshots on its own and typically reports them to a
`NativeTunnelController`, on a periodical basis
- The controller invokes a native C/JNI function to forward the
snapshots to the `PartoutVpnServiceRuntime.onSnapshots()` method
- The tunnel service broadcasts the snapshots for the app to eventually
observe them with a `BroadcastReceiver`

Lastly, change DataCount from UInt to UInt64 because it can overflow.

+++ cross/android/io/partout/abi/TunnelRemoteInfoWrapper.kt
@@ -35,7 +35,7 @@ import kotlinx.serialization.Contextual
 data class TunnelRemoteInfoWrapper (
 
     @SerialName(value = "fileDescriptors")
-    val fileDescriptors: kotlin.collections.List<kotlin.Int>,
+    val fileDescriptors: kotlin.collections.List<kotlin.Long>,
 
     @SerialName(value = "originalModuleId")
     val originalModuleId: kotlin.String,
